implements beaconcha.in API for landing page
reverts Infura beaconchain API changes
